+++ pypsa/io.py
@@ -16,8 +16,6 @@
 """Functions for importing and exporting data.
 """
 
-from six import iteritems, iterkeys, string_types
-
 __author__ = "Tom Brown (FIAS), Jonas Hoersch (FIAS)"
 __copyright__ = "Copyright 2015-2017 Tom Brown (FIAS), Jonas Hoersch (FIAS), GNU GPL 3"
 
@@ -81,7 +79,15 @@ class ImporterCSV(Importer):
     def get_snapshots(self):
         fn = os.path.join(self.csv_folder_name, "snapshots.csv")
         if not os.path.isfile(fn): return None
-        return pd.read_csv(fn, index_col=0, encoding=self.encoding, parse_dates=True)
+        df = pd.read_csv(fn, index_col=0, encoding=self.encoding, parse_dates=True)
+        if "snapshot" in df:
+            df["snapshot"] = pd.to_datetime(df.snapshot)
+        return df
+
+    def get_investment_periods(self):
+        fn = os.path.join(self.csv_folder_name, "investment_periods.csv")
+        if not os.path.isfile(fn): return None
+        return pd.read_csv(fn, index_col=0, encoding=self.encoding)
 
     def get_static(self, list_name):
         fn = os.path.join(self.csv_folder_name, list_name + ".csv")
@@ -117,6 +123,10 @@ class ExporterCSV(Exporter):
         fn = os.path.join(self.csv_folder_name, "snapshots.csv")
         snapshots.to_csv(fn, encoding=self.encoding)
 
+    def save_investment_periods(self, investment_periods):
+        fn = os.path.join(self.csv_folder_name, "investment_periods.csv")
+        investment_periods.to_csv(fn, encoding=self.encoding)
+
     def save_static(self, list_name, df):
         fn = os.path.join(self.csv_folder_name, list_name + ".csv")
         df.to_csv(fn, encoding=self.encoding)
@@ -146,6 +156,9 @@ class ImporterHDF5(Importer):
 
     def get_snapshots(self):
         return self.ds["/snapshots"] if "/snapshots" in self.ds else None
+
+    def get_investment_periods(self):
+        return self.ds["/investment_periods"] if "/investment_periods" in self.ds else None
 
     def get_static(self, list_name):
         if "/" + list_name not in self.ds:
@@ -182,6 +195,9 @@ class ExporterHDF5(Exporter):
     def save_snapshots(self, snapshots):
         self.ds.put('/snapshots', snapshots, format='table', index=False)
 
+    def save_investment_periods(self, investment_periods):
+        self.ds.put('/investment_periods', investment_periods, format='table', index=False)
+
     def save_static(self, list_name, df):
         df.index.name = 'name'
         self.index[list_name] = df.index
@@ -196,27 +212,30 @@ if has_xarray:
     class ImporterNetCDF(Importer):
         def __init__(self, path):
             self.path = path
-            if isinstance(path, string_types):
+            if isinstance(path, str):
                 self.ds = xr.open_dataset(path)
             else:
                 self.ds = path
 
         def __enter__(self):
-            if isinstance(self.path, string_types):
+            if isinstance(self.path, str):
                 super(ImporterNetCDF, self).__init__()
             return self
 
         def __exit__(self, exc_type, exc_val, exc_tb):
-            if isinstance(self.path, string_types):
+            if isinstance(self.path, str):
                 super(ImporterNetCDF, self).__exit__(exc_type, exc_val, exc_tb)
 
         def get_attributes(self):
             return {attr[len('network_'):]: val
-                    for attr, val in iteritems(self.ds.attrs)
+                    for attr, val in self.ds.attrs.items()
                     if attr.startswith('network_')}
 
         def get_snapshots(self):
             return self.get_static('snapshots', 'snapshots')
+
+        def get_investment_periods(self):
+            return self.get_static('investment_periods', 'investment_periods')
 
         def get_static(self, list_name, index_name=None):
             t = list_name + '_'
@@ -227,14 +246,14 @@ if has_xarray:
                 return None
             index = self.ds.coords[index_name].to_index().rename('name')
             df = pd.DataFrame(index=index)
-            for attr in iterkeys(self.ds.data_vars):
+            for attr in self.ds.data_vars.keys():
                 if attr.startswith(t) and attr[i:i+2] != 't_':
                     df[attr[i:]] = self.ds[attr].to_pandas()
             return df
 
         def get_series(self, list_name):
             t = list_name + '_t_'
-            for attr in iterkeys(self.ds.data_vars):
+            for attr in self.ds.data_vars.keys():
                 if attr.startswith(t):
                     df = self.ds[attr].to_pandas()
                     df.index.name = 'name'
@@ -249,12 +268,17 @@ if has_xarray:
 
         def save_attributes(self, attrs):
             self.ds.attrs.update(('network_' + attr, val)
-                                 for attr, val in iteritems(attrs))
+                                 for attr, val in attrs.items())
 
         def save_snapshots(self, snapshots):
             snapshots.index.name = 'snapshots'
             for attr in snapshots.columns:
                 self.ds['snapshots_' + attr] = snapshots[attr]
+
+        def save_investment_periods(self, investment_periods):
+            investment_periods.index.name = 'investment_periods'
+            for attr in investment_periods.columns:
+                self.ds['investment_periods_' + attr] = investment_periods[attr]
 
         def save_static(self, list_name, df):
             df.index.name = list_name + '_i'
@@ -297,7 +321,7 @@ def _export_to_exporter(network, exporter, basename, export_standard_types=False
 
     #exportable component types
     #what about None???? - nan is float?
-    allowed_types = (float,int,bool) + string_types + tuple(np.typeDict.values())
+    allowed_types = (float, int, bool, str) + tuple(np.typeDict.values())
 
     #first export network properties
     attrs = dict((attr, getattr(network, attr))
@@ -307,9 +331,17 @@ def _export_to_exporter(network, exporter, basename, export_standard_types=False
     exporter.save_attributes(attrs)
 
     #now export snapshots
-    snapshots = pd.DataFrame(dict(weightings=network.snapshot_weightings),
-                             index=pd.Index(network.snapshots, name="name"))
+    if isinstance(network.snapshot_weightings.index, pd.MultiIndex):
+        network.snapshot_weightings.index.rename(["period", "snapshot"], inplace=True)
+    else:
+        network.snapshot_weightings.index.rename("snapshot", inplace=True)
+    snapshots = network.snapshot_weightings.reset_index()
     exporter.save_snapshots(snapshots)
+
+    # export investment period weightings
+    # TODO: uncomment as soon as merge multi-decade branch
+    # investment_periods = network.investment_period_weightings
+    # exporter.save_investment_periods(investment_periods)
 
     exported_components = []
     for component in network.all_components - {"SubNetwork"}:
@@ -358,7 +390,7 @@ def _export_to_exporter(network, exporter, basename, export_standard_types=False
                     col_export = pnl[attr].columns[(pnl[attr] != default).any()]
 
             if len(col_export) > 0:
-                df = pnl[attr][col_export]
+                df = pnl[attr].reset_index()[col_export]
                 exporter.save_series(list_name, attr, df)
             else:
                 exporter.remove_series(list_name, attr)
@@ -492,7 +524,7 @@ def import_from_netcdf(network, path, skip_time=False):
 
     assert has_xarray, "xarray must be installed for netCDF support."
 
-    basename = os.path.basename(path) if isinstance(path, string_types) else None
+    basename = os.path.basename(path) if isinstance(path, str) else None
     with ImporterNetCDF(path=path) as importer:
         _import_from_importer(network, importer, basename=basename,
                               skip_time=skip_time)
@@ -565,7 +597,7 @@ def _import_from_importer(network, importer, basename, skip_time=False):
         except KeyError:
             pypsa_version = None
 
-        for attr, val in iteritems(attrs):
+        for attr, val in attrs.items():
             setattr(network, attr, val)
 
     ##https://docs.python.org/3/tutorial/datastructures.html#comparing-sequences-and-other-types
@@ -581,10 +613,31 @@ def _import_from_importer(network, importer, basename, skip_time=False):
 
     # if there is snapshots.csv, read in snapshot data
     df = importer.get_snapshots()
+    # read in investment period weightings
+    # TODO: uncomment as soon as merge multi-decade branch
+    # investment_periods = importer.get_investment_periods()
+
     if df is not None:
+
+        # check if imported snapshots have MultiIndex
+        snapshot_levels = set(["period", "snapshot"]).intersection(df.columns)
+        if snapshot_levels:
+            df.set_index(sorted(snapshot_levels), inplace=True)
         network.set_snapshots(df.index)
-        if "weightings" in df.columns:
+
+        cols = ['objective', 'generators', 'stores']
+        if not df.columns.intersection(cols).empty:
+            network.snapshot_weightings = df.reindex(index=network.snapshots,
+                                                     columns=cols)
+        elif "weightings" in df.columns:
             network.snapshot_weightings = df["weightings"].reindex(network.snapshots)
+
+        network.set_snapshots(df.index)
+
+        # TODO: uncomment as soon as merge multi-decade branch
+        # if investment_periods is not None:
+        #     network.investment_period_weightings = (
+        #         investment_periods.reindex(network.investment_period_weightings.index))
 
     imported_components = []
 
@@ -604,6 +657,7 @@ def _import_from_importer(network, importer, basename, skip_time=False):
 
         if not skip_time:
             for attr, df in importer.get_series(list_name):
+                df.set_index(network.snapshots, inplace=True)
                 import_series_from_dataframe(network, df, component, attr)
 
         logger.debug(getattr(network,list_name))
@@ -766,9 +820,9 @@ def import_series_from_dataframe(network, dataframe, cls_name, attr):
         dataframe = dataframe.reindex(network.snapshots, fill_value=default)
 
     if not attr_series.static:
-        pnl[attr] = pnl[attr].reindex(columns=df.index | columns, fill_value=default)
+        pnl[attr] = pnl[attr].reindex(columns=df.index.union(columns), fill_value=default)
     else:
-        pnl[attr] = pnl[attr].reindex(columns=(pnl[attr].columns | columns))
+        pnl[attr] = pnl[attr].reindex(columns=(pnl[attr].columns.union(columns)))
 
     pnl[attr].loc[network.snapshots, columns] = dataframe.loc[network.snapshots, columns]