+++ doc/release_notes.rst
@@ -15,6 +15,25 @@ Upcoming Release
 
 * Fix setting ``margin`` and ``boundaries`` when plotting a network with ``geomap=False``. 
 
+* Adjust log file creation for CPLEX version 12.10 and higher.
+
+* ``network.snapshots`` are now a property, hence assigning values with ``network.snapshots = values `` is the same as ``network.set_snapshots(values)`` 
+
+* ``network.snapshot_weightings`` are now subdivided into weightings for the objective function, generators and stores/storage units.
+
+  * Objective weightings determine the multiplier of the marginal costs in the objective function of the LOPF. 
+  * Generator weightings specify the impact of generators in a ``GlobalConstraint``. 
+  * Store weightings define the elapsed hours for the charge, discharge, standing loss and spillage of storage units and stores in order to determine the current state of charge.
+ 
+  PyPSA still supports setting ``snapshot_weightings`` with a ``pandas.Series``. In this case, the weightings are uniformly applied to all columns of the new ``snapshot_weightings`` ``pandas.DataFrame``.
+
+* The function ``geo.area_from_lon_lat_poly`` was deprecated and will be removed in v0.19.
+
+* All ``pypsa`` functionalities except for optimization with ``pyomo`` work now with multi-indexed snapshots.
+
+* A new module `examples` was added which contains frontend functions for retrieving/loading example networks provided by the PyPSA project. 
+
+
 PyPSA 0.17.1 (15th July 2020)
 =============================
 
@@ -187,7 +206,7 @@ package in conda is updated.
   according to the distribution scheme provided in the argument
   ``slack_weights``. If ``distribute_slack=False`` only the slack
   generator takes up the slack. There is further `documentation
-  <https://pypsa.readthedocs.io/en/latest/power_flow.html#full-non-linear-power-flow>`_.
+  <https://pypsa.readthedocs.io/en/latest/power_flow.html#full-non-linear-power-flow>`__.
 
 * Unit testing is now performed on all of GNU/Linux, Windows and MacOS.
 
@@ -272,7 +291,7 @@ This release contains a new feature and bug fixes.
   generators being handled correctly when networks are aggregated.
 * Network.consistency_check() now only prints necessary columns when
   reporting NaN values.
-* Import from `pandapower <https://www.pandapower.org/>`_ networks has
+* Import from `pandapower <https://www.pandapower.org/>`__ networks has
   been updated to pandapower 2.0 and to include non-standard lines and
   transformers.
 
@@ -593,7 +612,7 @@ changes to the internal API.
   :ref:`line-types`. For transformers you just need to specify the
   type, see :ref:`transformer-types`. The implementation of PyPSA's
   standard types is based on `pandapower's standard types
-  <http://www.uni-kassel.de/eecs/fileadmin/datas/fb16/Fachgebiete/energiemanagement/Software/pandapower-doc/std_types/basic.html>`_. The
+  <https://pandapower.readthedocs.io/en/latest/std_types/basic.html>`_. The
   old interface of specifying r, x, b and g manually is still available.
 * The transformer model has been substantially overhauled, see
   :ref:`transformer-model`. The equivalent model now defaults to the